feature : 매칭 리스트 조회 기능 구현

+++ match-service/src/main/java/com/ohdelivery/service/match/rider/application/dto/response/UpdateRiderStstusResponse.java
@@ -10,7 +10,7 @@ import lombok.RequiredArgsConstructor;
 public class UpdateRiderStstusResponse {
 
   private final UUID id;
-  private final Integer rider_id;
+  private final long rider_id;
   private final String slack_id;
   private final String status;
   private final Double latitude;
